requirements.txt
SQLAlchemy
python-dotenv
Flask
PyInquirer